fix: Let Profile header update on Relay data change
